Move internal config api to core module
