+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/config/DisabledAutoCorrectConfig.kt
@@ -1,7 +1,9 @@
-package io.gitlab.arturbosch.detekt.api.internal
+package io.gitlab.arturbosch.detekt.core.config
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.Notification
+import io.gitlab.arturbosch.detekt.api.internal.ValidatableConfiguration
+import io.gitlab.arturbosch.detekt.api.internal.validateConfig
 
 @Suppress("UNCHECKED_CAST")
 class DisabledAutoCorrectConfig(private val wrapped: Config) : Config, ValidatableConfiguration {
