+++ detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/config/YamlConfigSpec.kt
@@ -1,7 +1,10 @@
-package io.gitlab.arturbosch.detekt.api.internal
+@file:Suppress("detekt.MaxLineLength")
+
+package io.gitlab.arturbosch.detekt.core.config
 
 import io.github.detekt.test.utils.resourceAsPath
 import io.gitlab.arturbosch.detekt.api.Config
+import io.gitlab.arturbosch.detekt.api.internal.YamlConfig
 import io.gitlab.arturbosch.detekt.test.yamlConfig
 import io.gitlab.arturbosch.detekt.test.yamlConfigFromContent
 import org.assertj.core.api.Assertions.assertThat
@@ -13,7 +16,6 @@ import org.spekframework.spek2.style.specification.describe
 import org.yaml.snakeyaml.parser.ParserException
 import java.nio.file.Paths
 
-@Suppress("DEPRECATION")
 class YamlConfigSpec : Spek({
 
     describe("load yaml config") {
