+++ detekt-core/src/test/resources/composite-test.yml
