+++ detekt-core/src/test/resources/config_validation/baseline.yml
