+++ detekt-core/src/test/resources/config_validation/default-excluded-properties.yml
