+++ detekt-core/src/test/resources/config_validation/no-nested-config.yml
