+++ detekt-core/src/test/resources/config_validation/no-value.yml
