+++ detekt-core/src/test/resources/config_validation/other-nested-property-names.yml
