+++ detekt-core/src/test/resources/config_validation/other-ruleset-name.yml
