+++ detekt-core/src/test/resources/detekt.txt
