+++ detekt-core/src/test/resources/detekt.yml
