+++ detekt-core/src/test/resources/oneitem.yml
