+++ detekt-core/src/test/resources/wrong-property-type.yml
