+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/nested-configuration-expected.md
@@ -1,0 +1,14 @@
+---
+title: nestedConfigurationExpected -
+---
+//[detekt-api](../index.md)/[io.gitlab.arturbosch.detekt.api.internal](index.md)/[nestedConfigurationExpected](nested-configuration-expected.md)
+
+
+
+# nestedConfigurationExpected  
+[jvm]  
+Content  
+fun [nestedConfigurationExpected](nested-configuration-expected.md)(prop: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)): [Notification](../io.gitlab.arturbosch.detekt.api/-notification/index.md)  
+
+
+
