+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/property-does-not-exists.md
@@ -1,0 +1,14 @@
+---
+title: propertyDoesNotExists -
+---
+//[detekt-api](../index.md)/[io.gitlab.arturbosch.detekt.api.internal](index.md)/[propertyDoesNotExists](property-does-not-exists.md)
+
+
+
+# propertyDoesNotExists  
+[jvm]  
+Content  
+fun [propertyDoesNotExists](property-does-not-exists.md)(prop: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)): [Notification](../io.gitlab.arturbosch.detekt.api/-notification/index.md)  
+
+
+
