+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/property-is-deprecated.md
@@ -1,0 +1,14 @@
+---
+title: propertyIsDeprecated -
+---
+//[detekt-api](../index.md)/[io.gitlab.arturbosch.detekt.api.internal](index.md)/[propertyIsDeprecated](property-is-deprecated.md)
+
+
+
+# propertyIsDeprecated  
+[jvm]  
+Content  
+fun [propertyIsDeprecated](property-is-deprecated.md)(prop: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html), deprecationDescription: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html), reportAsError: [Boolean](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-boolean/index.html)): [Notification](../io.gitlab.arturbosch.detekt.api/-notification/index.md)  
+
+
+
