+++ docs/pages/kdoc/detekt-api/io.gitlab.arturbosch.detekt.api.internal/unexpected-nested-configuration.md
@@ -1,0 +1,14 @@
+---
+title: unexpectedNestedConfiguration -
+---
+//[detekt-api](../index.md)/[io.gitlab.arturbosch.detekt.api.internal](index.md)/[unexpectedNestedConfiguration](unexpected-nested-configuration.md)
+
+
+
+# unexpectedNestedConfiguration  
+[jvm]  
+Content  
+fun [unexpectedNestedConfiguration](unexpected-nested-configuration.md)(prop: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)): [Notification](../io.gitlab.arturbosch.detekt.api/-notification/index.md)  
+
+
+
